test: extract InjectableRemoteScanner base shared across test files
